Use icons for Activate/Deactive/Settings links
An unchecked checkbox icon is used for Activate (to show that it's
currently inactive). A checked checkbox icon is used for Deactivate (to
show that it's currently active). A gear icon is used for the Settings
link.

The checkbox icons are placed on the left side of the name, as checkboxes
usually appear. The gear icon is placed floated to the right side of the
name, to have a clean look.

The plugin list is more compact, and it's much easier to identify which
plugins are currently active, and which have settings. No need to read
any words!

The previously used words (with language dependency) is still used as the
tooltip (title) that is shown when the mouse hovers over the icons.

In the themes page, the checked checkbox icon denotes the curently active
theme, but doesn't do anything when clicked. This is by design, since only
one theme can be active at a time. For consistency sake, a circle (radio
button style) was not used.

Signed-off-by: ADTC <send2adtc@gmail.com>