Generate title (if empty) from the content.
A new function is created to achieve this, Text::truncate
which is based on Text::cut from the text.class.php helper.

Limit is set to 60 characters.

Example A:
1. Content has over 60 characters.
2. Title = First 60 characters + remaining characters of the last word + '...'

Example B:
1. Content has less than or equal to 60 characters
2. Title = All.

Example C:
1. Content only has one word but is over 60 characters.
2. Title = First 60 characters + '...'

Why B and C differ from each other?
Because we do not want the whole that one weird word
with over 60 characters to make its way to the title.
It's probably unrealistic, the code can be much cleaner
with B and C combined.

Feel free to discuss.